+++ packages/sdk/src/identity/IdentityClient.ts
@@ -2,6 +2,7 @@ import {
   DEFAULT_IDENTITY_CLIENT_OPTIONS,
   defaultIdentity,
   DisplayableIdentity,
+  IdentityClientOptions,
   KNOWN_IDENTITY_TYPES
 } from './types/index.js'
 import {
@@ -102,11 +103,13 @@ function normalizeOpts(
 export class IdentityClient {
   private readonly wallet: WalletInterface
   private readonly contactsManager: ContactsManager
+  private readonly options: IdentityClientOptions
   constructor(
     wallet?: WalletInterface,
-    private readonly options = DEFAULT_IDENTITY_CLIENT_OPTIONS,
+    options: Partial<IdentityClientOptions> = {},
     private readonly originator?: OriginatorDomainNameStringUnder250Bytes
   ) {
+    this.options = { ...DEFAULT_IDENTITY_CLIENT_OPTIONS, ...options }
     this.wallet = wallet ?? new WalletClient()
     this.contactsManager = new ContactsManager(this.wallet, this.originator)
   }